Move LeaseFileTabs into the leases folder
