+++ source/frontend/src/features/properties/map/lease/ViewSelector.tsx
@@ -1,5 +1,5 @@
 import { LeaseFormModel } from 'features/leases/models';
-import { LeaseFileTabNames } from 'features/mapSideBar/tabs/LeaseFileTabs';
+import { LeaseFileTabNames } from 'features/properties/map/lease/detail/LeaseFileTabs';
 import { FormikProps } from 'formik';
 import { IFormLease, ILease } from 'interfaces';
 import * as React from 'react';
