+++ source/frontend/src/features/properties/map/lease/detail/LeaseFileTabs.tsx
